move parseContent from support to CMS folder
